RC/NavBar - Bring back Search Key/Action [1/2]
I had removed 'Search' as an option for the NavBar.
My thought was that we would use Search as a shortcut/app.
It didn't occur to me that the search 'key' could be useful for
in app searching for those that supported it.
This will put the option back in RC. I will also push
a patch for the frameworks side.

Change-Id: I32ce71a9f607e09251eb3ef5101b310d21591b1c
Signed-off-by: Zaphod <mwilson@acesappraisal.com>